DROP xlsx update
Created a new version of the script intended for the DROP facility. This new version includes a .lvm to .xlsx converter which also trims the data to the fundamental variables. Afterwards, the script works with this .xlsx file the same way the old one did with .lvm files. The old script is preserved but its further development will be discontinued.